suit/requirements.txt
blinker==1.8.2
cffi==1.17.1
click==8.1.7
cryptography==43.0.1
Flask==3.0.3
Flask-Cors==5.0.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.2
pycparser==2.22
PyMySQL==1.1.1
python-dotenv==1.0.1
Werkzeug==3.0.4
